updated component names part 2
