+++ src/components/secretary/module/secretary-week-all.js
@@ -3,8 +3,9 @@ import React, {Component} from 'react';
 import {Tab, Table} from 'semantic-ui-react';
 import AppointmentCard from './secretary-edit-appointment-modal'
 import moment from 'moment'
+import qs from 'qs'
 
-class day_all extends Component{
+class week_all extends Component{
 
     constructor(props){
         super(props);
@@ -13,17 +14,17 @@ class day_all extends Component{
             appointments:[],
             currNumOfApp: 0,
             prevNumOfApp: 0,
-            day: this.props.day,
+            weeks: this.props.week,
         }
 
     }
     
 
     componentDidMount(){
-       const day = {
-           day: this.state.day
+       const week = {
+           weeks: this.state.weeks
        }
-        Axios.post('http://localhost:3000/secretary/day_all', day).then(res =>{
+        Axios.post('http://localhost:3000/secretary/week_all', week).then(res =>{
             return this.setState({
                 appointments: res.data.data.data
             })
@@ -96,4 +97,4 @@ class day_all extends Component{
     }
 }
 
-export default day_all
+export default week_all
